wr-common-packages-native.bb: put host-tools' sstate files in the top level
Issue: LIN7-564

The native sstate cache doesn't allow to use between distributions, for
example, Redhat and Ubuntu, but we need it. All the native's sstate
cache files were put in the distro related directory, for example, in
sstate-cache/Ubuntu-13.10 in the past, but now, the populate-lic is in
sstate-cache/, others are in sstate-cache/Ubuntu-13.10, we can move all
of them to the top dir to make it work.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>