lttng-tools: enable python for WR and move it to bbappend
The oe-core disable python-bindings by default, because it
requires swig-native as DEPENDS, and swig-native is not
in oe-core.

But windriver have swig in meta-oe-subset, so we enable
python-bindings for wr specific and previously we added
this in oe-core as LOCAL REV, we move this back into
bbappend here.

Issue: TA61857

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>