rsnapshot: fix rsnapshot.conf.default
Issue: LIN7-1020

This change is used for fixing cmd path in rsnapshot.conf.default.

The options --without-* disable checking command * on host and get
the default path used, otherwise the host path will be injected into
target configs.

The runtime dependencies to ssh, logger, cp, du are optional and
could be customized in rsnapshot.conf, so it's not needed that
using PACKAGECONFIG to define the runtime dependencies.

Signed-off-by: Wenzong Fan <wenzong.fan@windriver.com>