oprofile: add 'sysvinit-pidof' to RDEPENDS
Issue: LIN7-1259

`opcontrol --shutdown' in systemd systems would result in error complaining
that `pidof' is not available.

In Yocto, the pidof command is provided by busybox, thus no such problem.

In WRLinux, we need to add sysvinit-pidof to RDEPENDS if the distro is
'wrlinux' or 'wrlinux-std-sato', because busybox for these two distros
doesn't provide the `pidof' command.

As this is a wrlinux specific fix, we put it in wr-base layer.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>