ttf-vlgothic: set LICENSE_FLAGS to commercial
Issue: LIN7-1422

fix fetching source code failure

The lisence of ttf-vlgothic is a commercial lisence. We just add a bb file for it
in wr-base layer. And its source code could not be provided by WR linux because of its
commercial lisence.

By default, the component should not be built because its commercial lisence.
The web download should be enable to get source code if customer want to use it.

Signed-off-by: Shiqun Lin <Shiqun.Lin@windriver.com>