ltp: fix .debug packaging
Issue: LIN7-3082, LIN7-3178

ltp never really generated any .debug directories because the
INHIBIT_PACKAGE_STRIP in the recipe suppressed their generation as
a side-effect.  This will no longer be true, because package.bbclass
is being fixed to make stripping and -dbg package population independent.
So the correct locations of unusual .debug directories need to be specified.
There are lots of potential directories, so we just search for all possible
ones.

Although it is not a fatal error, we also avoid a qa error stating
that test_arch_stripped is stripped.

Signed-off-by: Joe Slater <jslater@windriver.com>