image: Condational assign kernel-modules to IMAGE_INSTALL
In order to support container image building, defined a global variable
IMAGE_ENABLE_CONTAINER. When a project is configured with --enable-container=yes,
IMAGE_ENABLE_CONTAINER will be set to "1", otherwise is set to "0".
Since the kernel related files are not necessary for container image,
change kernel-module to conditionally assigned to IMAGE_INSTALL based on
IMAGE_ENABLE_CONTAINER's value. When IMAGE_ENABLE_CONTAINER="1", "kernel-module"
will be appended to IMAGE_INSTALL, otherwise, it will append nothing to it.

Signed-off-by: fupan li <fupan.li@windriver.com>