packagegroup-core-tools-testapps: remove duplicate recipe
Issue: TA121602

Remove duplicate recipe of packagegroup-core-tools-testapps in wr-base
and use the one in oe-core. It is from oe-core originally and almost
same with the recipe in oe-core. Most of the differences are caused by
update in oe-core.

Only need to push back set of KEXECTOOLS for mips64 and mips64n32.

Signed-off-by: Kai Kang <kai.kang@windriver.com>